+++ includes/class-wdevs-tax-switch-block-switch.php
@@ -40,7 +40,9 @@ class Wdevs_Tax_Switch_Block_Switch extends Wdevs_Tax_Switch_Block {
 	 * @since 1.0,0
 	 */
 	public function init_block() {
+		$editor_asset = $this->register_script('wdevs-tax-switch-editor-script', 'switch', 'index');
 		register_block_type( plugin_dir_path( dirname( __FILE__ ) ) . 'build/switch/block.json', array(
+			'editor_script'   => 'wdevs-tax-switch-editor-script',
 			'render_callback' => [ $this, 'block_render_callback' ],
 		) );
 
@@ -105,10 +107,6 @@ class Wdevs_Tax_Switch_Block_Switch extends Wdevs_Tax_Switch_Block {
 	public function enqueue_frontend_scripts() {
 		if ( wp_style_is( 'wdevs-tax-switch-style', 'registered' ) ) {
 			wp_enqueue_style( 'wdevs-tax-switch-style' );
-		}
-
-		if ( wp_script_is( 'wdevs-tax-switch-shared-script', 'registered' ) ) {
-			wp_enqueue_script( 'wdevs-tax-switch-shared-script' );
 		}
 
 		if ( wp_script_is( 'wdevs-tax-switch-view-script', 'registered' ) ) {