+++ webpack.config.js
@@ -44,11 +44,6 @@ module.exports = {
 			'block/src/switch',
 			'woocommerce-quantity-manager.js'
 		),
-		'switch/variation-price-display-range-for-wc': path.resolve(
-			process.cwd(),
-			'block/src/switch',
-			'variation-price-display-range-for-wc.js'
-		),
 		'label/index': path.resolve(
 			process.cwd(),
 			'block/src/label',
@@ -74,7 +69,6 @@ module.exports = {
 					enforce: true,
 					test: /[\\/]block[\\/]src[\\/]/,
 				},
-
 			},
 		},
 	},