+++ CHANGELOG.md
@@ -1,12 +1,18 @@
 # Changelog
 All notable changes to the Tax Switch for WooCommerce plugin will be documented in this file.
 
+## [1.5.4] - 2025-05-26
+### Added
+- Introduced 'tree shake' for bundling all commonly shared JavaScript into a separate file as optimisation to prevent duplicated code
+### Updated
+- Improved compatibility for Variation Price Display Range for WooCommerce
+
 ## [1.5.3] - 2025-05-21
 ### Added
 - Compatibility for Variation Swatches for WooCommerce
 - Preview in shortcode generator
 - JavaScript 'wdevs-tax-switch-appeared' event listener
-- Compatibility for Variation Price Display Range for WooCommerce
+- Compatibility for Variation Price Display Range for WooCommerce (1.3.20 and up)
 - Added settings link
 ### Updated
 - Switched from 'found_variation' to 'show_variation' WooCommerce event for UI updates reliability in variations form